test(no_debt_gate): split F34 allowlist tests into sibling file (300-line cap)
