+++ crates/convergio-durability/tests/no_debt_gate.rs
@@ -229,114 +229,3 @@ async fn fires_through_full_facade_pipeline() {
     assert!(msg.contains("no_debt"), "msg: {msg}");
     assert!(msg.contains("rust_todo_macro"), "msg: {msg}");
 }
-
-// ---------- F34: debt-topic allowlist ----------
-//
-// A task whose title declares it is *about* a debt-related artifact
-// (WIP-commit template, TODO inventory, debt burn-down spec) cannot
-// avoid the keyword in its documentation evidence. The gate skips
-// the marker scan when BOTH the evidence kind is documentation/spec
-// AND the task title contains a debt-topic token. Code/test evidence
-// on the same task is still scanned.
-
-async fn make_titled_task(
-    dur: &Durability,
-    title: &str,
-    evidence: Vec<(&str, serde_json::Value)>,
-) -> convergio_durability::Task {
-    let plan = dur
-        .create_plan(NewPlan {
-            title: "p".into(),
-            description: None,
-            project: None,
-        })
-        .await
-        .unwrap();
-    let task = dur
-        .create_task(
-            &plan.id,
-            NewTask {
-                wave: 1,
-                sequence: 1,
-                title: title.into(),
-                description: None,
-                evidence_required: vec![],
-            },
-        )
-        .await
-        .unwrap();
-    for (kind, payload) in evidence {
-        dur.attach_evidence(&task.id, kind, payload, Some(0))
-            .await
-            .unwrap();
-    }
-    dur.tasks().get(&task.id).await.unwrap()
-}
-
-#[tokio::test]
-async fn allows_wip_in_doc_on_wip_titled_task() {
-    let (dur, _dir) = fresh().await;
-    let task = make_titled_task(
-        &dur,
-        "WIP commit message protocol",
-        vec![(
-            "doc",
-            json!({"summary": "canonical WIP-commit pause-checklist"}),
-        )],
-    )
-    .await;
-    NoDebtGate::default()
-        .check(&ctx(&dur, task, TaskStatus::Submitted))
-        .await
-        .expect("WIP in doc on WIP-titled task should be allowed (F34)");
-}
-
-#[tokio::test]
-async fn still_refuses_unwrap_in_code_on_wip_titled_task() {
-    let (dur, _dir) = fresh().await;
-    let task = make_titled_task(
-        &dur,
-        "WIP commit message protocol",
-        vec![("code", json!({"diff": "let x = parse(s).unwrap();"}))],
-    )
-    .await;
-    let err = NoDebtGate::default()
-        .check(&ctx(&dur, task, TaskStatus::Submitted))
-        .await
-        .unwrap_err();
-    assert!(err.to_string().contains("rust_unwrap"));
-}
-
-#[tokio::test]
-async fn refuses_wip_in_doc_on_unrelated_task_title() {
-    let (dur, _dir) = fresh().await;
-    let task = make_titled_task(
-        &dur,
-        "Add user-facing dashboard",
-        vec![("doc", json!({"summary": "WIP draft, leaving notes"}))],
-    )
-    .await;
-    let err = NoDebtGate::default()
-        .check(&ctx(&dur, task, TaskStatus::Submitted))
-        .await
-        .unwrap_err();
-    assert!(err.to_string().contains("todo_marker"));
-}
-
-#[tokio::test]
-async fn allowlist_covers_kind_spec_too() {
-    let (dur, _dir) = fresh().await;
-    let task = make_titled_task(
-        &dur,
-        "Debt burn-down spec",
-        vec![(
-            "spec",
-            json!({"text": "We track FIXME and TODO across the repo."}),
-        )],
-    )
-    .await;
-    NoDebtGate::default()
-        .check(&ctx(&dur, task, TaskStatus::Submitted))
-        .await
-        .expect("FIXME/TODO mention in spec on debt-titled task should be allowed (F34)");
-}
